Modularize negotiation buffer helpers
